added another custom component for header background
