+++ components/main-header/main-header.jsx
@@ -1,0 +1,34 @@
+import Link from "next/link";
+import Image from "next/image";
+
+import React from "react";
+import logoImg from "@/assets/logo.png";
+import classes from "./main-header.module.css";
+import MainHeaderBackground from "./main-header-background";
+
+const MainHeader = () => {
+  return (
+    <>
+      <MainHeaderBackground />
+      <header className={classes.header}>
+        <Link className={classes.logo} href="/">
+          <Image src={logoImg} alt="A plate with food on it" priority />
+          Next-Food
+        </Link>
+
+        <nav className={classes.nav}>
+          <ul>
+            <li>
+              <Link href="/meals">Browse Meals</Link>
+            </li>
+            <li>
+              <Link href="/community">Foodies Community</Link>
+            </li>
+          </ul>
+        </nav>
+      </header>
+    </>
+  );
+};
+
+export default MainHeader;
